perf: simplify sumtree implementation
